Moved the event setup in fixture
